chore: Split up LocationAutocompleteInputs
