+++ src/app/components/client/ComboBox.module.scss
@@ -1,6 +1,6 @@
 @import "../../tokens";
 
-.locationInput {
+.comboBox {
   background: $color-white;
   display: flex;
   flex-direction: column;
